move logging to its own file
