+++ src/utils.js
@@ -2,28 +2,14 @@ const { spawn } = require('child_process')
 const { relative } = require('path')
 const { existsSync, writeFile, unlink } = require('fs')
 const { promisify } = require('util')
-
-const formatCyan = '\x1b[36m'
-const formatGreen = '\x1b[32m'
-const formatYellow = '\x1b[33m'
-const formatRed = '\x1b[31m'
-const formatReset = '\x1b[0m'
-const checkmark = `${formatGreen}✔︎${formatReset}`
-
-function logInfo(...args) {
-  console.info(`${formatCyan}[sammie]${formatReset}`, ...args) // eslint-disable-line no-console
-}
-
-function logCommand(...args) {
-  console.log(formatYellow, ...args, formatReset) // eslint-disable-line no-console
-}
+const { logError } = require('./log')
 
 async function spawnAsync(command) {
   const child = spawn(command, { shell: true })
   return new Promise((resolve, reject) => {
     let data = ''
     child.stdout.on('data', chunk => (data += chunk))
-    child.stderr.on('data', data => console.error(formatRed, data.toString(), formatReset)) // eslint-disable-line no-console
+    child.stderr.on('data', data => logError(data.toString()))
     child.on('error', reject)
     child.on('exit', code => {
       let response
@@ -55,17 +41,11 @@ function findTemplatePath(input) {
   throw Error('Template not found')
 }
 
-const writeFileAsync = promisify(writeFile)
-const delteFileAsync = promisify(unlink)
-
 module.exports = {
-  logInfo,
-  logCommand,
-  checkmark,
   spawnAsync,
   stackSafeName,
   resourceSafeName,
   findTemplatePath,
-  writeFileAsync,
-  delteFileAsync
+  writeFileAsync: promisify(writeFile),
+  delteFileAsync: promisify(unlink)
 }
